+++ packages/umbrella/CHANGELOG.md
@@ -1,5 +1,14 @@
 # glissade
 
+## 0.55.0-pre.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.55.0-pre.2
+  - @glissade/player@0.55.0-pre.2
+  - @glissade/core@0.55.0-pre.2
+
 ## 0.55.0-pre.1
 
 ### Patch Changes